+++ 2026/BS/tools/shared/src/bsSearch.kt
@@ -6,25 +6,7 @@ import io.ktor.client.request.*
 import io.ktor.client.statement.*
 import io.ktor.http.*
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.collectIndexed
 import kotlinx.coroutines.flow.flow
-
-
-//if (args.getOrNull(0) == "search") runBlocking {
-//    bsSearch(
-//        keywords = args.drop(1).joinToString(" ")
-//    ).collectIndexed { index, card -> println("$index: $card") }
-//}
-//
-//suspend fun bsSearch(
-//    keywords: String,
-//    cardNo: String = "",
-//    costMin: Int = 0,
-//    costMax: Int = 99,
-//    attr: String = "", // 全色の場合 "赤紫緑白黄青"
-//    category: List<String> = emptyList(),
-//    system: List<String> = emptyList(),
-//): Flow<SearchCard> = bsSearchMain(keywords, cardNo, costMin, costMax, attr, category, system)
 
 suspend fun bsSearchMain(
     keywords: String,