+++ 2026/BS/tools/shared/src@mingw/Funcs.kt
@@ -1,10 +1,8 @@
 import io.ktor.client.HttpClient
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.serialization.kotlinx.json.json
 import kotlinx.cinterop.ExperimentalForeignApi
 import kotlinx.cinterop.toKString
 import platform.posix.getenv
 
 @OptIn(ExperimentalForeignApi::class)
 actual fun getEnv(key: String): String = getenv(key)?.toKString() ?: ""
-actual val client = HttpClient { install(ContentNegotiation) { json() } }
+actual val client = HttpClient ()